fastapi/requirements.txt
fastapi==0.103.1
uvicorn==0.23.1
sqlalchemy==2.0.20
psycopg2-binary==2.9.7

annotated-types==0.5.0
anyio==3.7.1
click==8.1.7
exceptiongroup==1.1.3
greenlet==2.0.2
h11==0.14.0
httptools==0.6.0
idna==3.4
pydantic==2.3.0
pydantic_core==2.6.3
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3

django-environ==0.9.0
google-cloud-secret-manager==2.15.1
google-cloud-storage==2.7.0
google-api-core==2.11.0
google-api-python-client==2.77.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0

python-dotenv==0.21.1

